Library Reference

SemanticModels.SemanticModels — Module

SemanticModels

provides the AbstractModel type and constructors for building metamodeling tooling for categorical model representations

SemanticModels.AbstractModel — Type

AbstractModel

a placeholder struct to dispatch on how to parse the expression tree into a model.

SemanticModels.model — Method

model(::AbstractModel, x) dig into the expression that describes a model and break it down into components. This allows you to construct a structured representation of the modeling problem at the expression level. Just like how julia modeling frameworks build structured representations of the problems in data structures. This version builds them at the expression level. The first argument is the type you want to construct, the second argument is the model structure that you want to analyze. For example

model(PetriModel, x::Petri.Model)::PetriModel

Category Theory

The following Category Theory operations ship with SemanticModels, you can use them as templates for defining your own model classes.

SemanticModels.CategoryTheory.AbstractCospan — Type

AbstractCospan

an abstract type for representing cospans. The essential API for subtypes of AbstractCospan are

left(c)::M

right(c)::M

pullback(c)::S

where M is the type of morphisms in the cospan, and S is the type of span that solves the pullback defined by c. See Cospan for an example.

SemanticModels.CategoryTheory.AbstractMorph — Type

AbstractMorph

an abstract type for representing morphisms. The essential API for subtypes of AbstractMorph are

dom(m)::T

codom(m)::T

func(m)::Function

where T is the type of the objects in the category. See FinSetMorph for an example.

SemanticModels.CategoryTheory.AbstractSpan — Type

AbstractSpan

an abstract type for representing spans. The essential API for subtypes of AbstractSpan are

left(s)::M

right(s)::M

pushout(s)::C

where M is the type of morphisms in the span, and C is the type of cospan that solves a pushout of span s. See Span for an example.

SemanticModels.CategoryTheory.Cospan — Type

Cospan{F,G} <: AbstractCospan

a general cospan type where types F and G are types of morphisms in the cospan

SemanticModels.CategoryTheory.Decorated — Type

Decorated{M,T}

a decoration applied to the objects of a morphism, where M is a type of morphism and type T is the category of the decoration

SemanticModels.CategoryTheory.FinSetMorph — Type

FinSetMorph{T,F}

morphisms in the category of Finite Sets. The objects are of type UnitRange{Int}. func(m::FinSetMorph) is a function that takes Int -> Int. FinSetMorphs can be constructed from a list of numbers. For example, FinSetMorph([1,3,2,3]) is the morphism that takes 1->1, 2->3, 3->2, 4->3 on domain 1:4 with codomain 1:3. When you define a morphism from a list of integers, the codomain is inferred from the largest element of the list. The domain must always be the 1:l where l is the length of the input list.

(f::FinSetMorph)(g::G) where G <: AbstractGraph

lift a finite set morphism (list of integers) to a graph homomorphism by its action on the vertex set. The graph h = f(g) is defined by taking the edges of g and relabeling their src and dst according to the function of f.

This method computes a valid graph homomorphism by definition.

SemanticModels.CategoryTheory.Span — Type

Span{F,G} <: AbstractSpan

a general span type where types F and G are types of morphisms in the span

SemanticModels.CategoryTheory.:⊔ — Method

⊔(f::FinSetMorph, g::FinSetMorph)

the union of two morphisms in a finite set.

SemanticModels.CategoryTheory.pushout — Method

pushout(s::Span{T, T}) where T <: Decorated

treat f,g as a decorated span and compute the pushout that is, the cospan of f=(f⊔g) and g=(a⊔b), with the decoration of (f⊔g)(d)

SemanticModels.CategoryTheory.pushout — Method

pushout(s::Span{T,T}) where T <: FinSetMorph

treat f,g as a span and compute the pushout that is, the cospan of f=(f⊔g) and g=(a⊔b)

SemanticModels.CategoryTheory.undecorate — Method

undecorate(s::Copan{T,T}) where T <: Decorated

remove decorations of a cospan of decorated morphisms

SemanticModels.CategoryTheory.undecorate — Method

undecorate(s::Span{T,T}) where T <: Decorated

remove decorations of a span of decorated morphisms

Model Classes

The following model class ship with SemanticModels, you can use them as templates for defining your own model classes.

SemanticModels.WiringDiagrams.canonical — Method

canonical(Syntax::Module, hom)

canonicalizes a hom expression by converting it to a WiringDiagram and then back again. The Syntax parameter is the Doctrine for the morphism such as FreeSymmetricMonoidalCategory.

SemanticModels.WiringDiagrams.drawhom — Function

drawhom(hom, name::String, format="svg")

draw a hom expression as a wiring diagram and store it as a file. Defaults to SVG format. The filename is <name.format>.

SemanticModels.WiringDiagrams.odeTemplate — Method

odeTemplate(d::WiringDiagram)

create an expression that defines a code that solves the ODE. Given just the wiring diagram, we don't know the paramters, initial conditions, or timedomain, so they are passed in as arguments to the function we generate.

These parameters and initial conditions are destructured in the main function so you can see what the code is expecting to receive by reading the generated output.

The structure of the timedomain is not implied by the wiring diagram so it is passed directly to the ODEProblem constructor. Any keyword arguments you pass to main() are forwarded to solve().

SemanticModels.WiringDiagrams.oderhs — Method

oderhs(d::WiringDiagram)

convert a wiring diagram into a dynamical system described as julia Exprs. We need to keep track of the names we give the variables and the parameters so those are the second and third arguments respectively.

Returns an expression that defines a function, a list of symbols representing the variables, and a list of symbols representing the parameter names.

see also fluxes(d::WiringDiagram) for implementation details

SemanticModels.PetriModels.dropdown — Method

dropdown(l::Model, c::Model, l′::Model)

compute c′ given l, c, l′, the formula for petrinets is

T{c'} = T{l'} \setdiff f(Tl) \cup a(Tc)

SemanticModels.CategoryTheory.pushout — Method

pushout(m1::Model, m2::Model)

compute the CT pushout of two models.

Catlab.Graphics.Graphviz.Graph — Method

Graph(f::OpenModel)

convert an OpenModel into a GraphViz Graph. This calls Graph(::Model) and then adds vertices (and edges) for the domain and codomain of the open model.

Index

SemanticModels.SemanticModels

Catlab.Graphics.Graphviz.Graph

SemanticModels.AbstractModel

SemanticModels.CategoryTheory.AbstractCospan

SemanticModels.CategoryTheory.AbstractMorph

SemanticModels.CategoryTheory.AbstractSpan

SemanticModels.CategoryTheory.Cospan

SemanticModels.CategoryTheory.Decorated

SemanticModels.CategoryTheory.FinSetMorph

SemanticModels.CategoryTheory.Span

SemanticModels.ExprModels.Parsers.AbstractCollector

SemanticModels.ExprModels.Parsers.FuncCollector

SemanticModels.ExprModels.Parsers.MetaCollector

SemanticModels.CategoryTheory.:⊔

SemanticModels.CategoryTheory.pushout

SemanticModels.CategoryTheory.pushout

SemanticModels.CategoryTheory.pushout

SemanticModels.CategoryTheory.undecorate

SemanticModels.CategoryTheory.undecorate

SemanticModels.ExprModels.Parsers.defs

SemanticModels.ExprModels.Parsers.findassign

SemanticModels.ExprModels.Parsers.findfunc

SemanticModels.ExprModels.Parsers.funcs

SemanticModels.ExprModels.Parsers.inexpr

SemanticModels.ExprModels.Parsers.parsefile

SemanticModels.ExprModels.Parsers.postwalk

SemanticModels.ExprModels.Parsers.prewalk

SemanticModels.PetriModels.dropdown

SemanticModels.WiringDiagrams.canonical

SemanticModels.WiringDiagrams.drawhom

SemanticModels.WiringDiagrams.odeTemplate

SemanticModels.WiringDiagrams.oderhs

SemanticModels.model