Library Reference

Library Reference

SemanticModels.SemanticModels — Module.

SemanticModels

provides the AbstractModel type and constructors for building hierarchical model representations.

source

SemanticModels.CombinedModel — Type.

CombinedModel

represents a model that has a fixed set of dependencies. It is the basic building block of a model DAG.

source

SemanticModels.BasicSIR — Method.

BasicSIR

defines a representation of Susceptible Infected Recovered models using Unitful numbers and expressions.

source

SemanticModels.solve — Method.

solve(m::AbstractModel)

executes the solving of a model for models with dependencies, those deps are executed first and then the node is solved. This is the function that evaluates the model DAG.

source

Index

SemanticModels.SemanticModels

SemanticModels.CombinedModel

SemanticModels.Dubstep.GraftCtx

SemanticModels.Dubstep.LPCtx

SemanticModels.Dubstep.TraceCtx

SemanticModels.Dubstep.TracedRun

SemanticModels.ModelTools.AbstractModel

SemanticModels.ModelTools.ExpODEModels.ExpODEModel

SemanticModels.ModelTools.ExpStateModels.ExpStateModel

SemanticModels.ModelTools.ExpStateModels.ExpStateTransition

SemanticModels.ModelTools.SimpleModels.SimpleModel

SemanticModels.ModelTools.Transformations.Pow

SemanticModels.Parsers.AbstractCollector

SemanticModels.Parsers.FuncCollector

SemanticModels.Parsers.MetaCollector

Base.:^

Base.getindex

Base.put!

Base.replace!

Base.setindex!

SemanticModels.BasicSIR

SemanticModels.Dubstep.replacefunc

SemanticModels.Dubstep.replacenorm

SemanticModels.Dubstep.trace

SemanticModels.ModelTools.and

SemanticModels.ModelTools.argslist

SemanticModels.ModelTools.bodyblock

SemanticModels.ModelTools.callsites

SemanticModels.ModelTools.funclines

SemanticModels.ModelTools.funcname

SemanticModels.ModelTools.head

SemanticModels.ModelTools.isblock

SemanticModels.ModelTools.iscall

SemanticModels.ModelTools.isexpr

SemanticModels.ModelTools.isfunc

SemanticModels.ModelTools.isimport

SemanticModels.ModelTools.issome

SemanticModels.ModelTools.isusing

SemanticModels.ModelTools.model

SemanticModels.ModelTools.model

SemanticModels.ModelTools.or

SemanticModels.ModelTools.pusharg!

SemanticModels.ModelTools.setarg!

SemanticModels.ModelTools.structured

SemanticModels.ModelTools.typegraph

SemanticModels.Parsers.defs

SemanticModels.Parsers.findassign

SemanticModels.Parsers.findfunc

SemanticModels.Parsers.funcs

SemanticModels.Parsers.inexpr

SemanticModels.Parsers.parsefile

SemanticModels.Parsers.postwalk

SemanticModels.Parsers.prewalk

SemanticModels.solve

SemanticModels.ModelTools.@typegraph