Library Reference

Library Reference

SemanticModels.SemanticModels — Module.

SemanticModels

provides the AbstractModel type and constructors for building hierarchical model representations.

source

SemanticModels.CombinedModel — Type.

CombinedModel

represents a model that has a fixed set of dependencies. It is the basic building block of a model DAG.

source

SemanticModels.SpringModel — Type.

SpringModel

represents the second order linear ODE goverened by hookeslaw.

source

DiffEqBase.solve — Method.

solve(m::AbstractModel)

executes the solving of a model for models with dependencies, those deps are executed first and then the node is solved. This is the function that evaluates the model DAG.

source

SemanticModels.BasicSIR — Method.

BasicSIR

defines a representation of Susceptible Infected Recovered models using Unitful numbers and expressions.

source

SemanticModels.hookeslaw — Method.

hookeslaw

the ODE representation of a spring. The solutions are periodic functions.

source

Index

SemanticModels.SemanticModels

SemanticModels.CombinedModel

SemanticModels.Dubstep.GraftCtx

SemanticModels.Dubstep.LPCtx

SemanticModels.Dubstep.TraceCtx

SemanticModels.Dubstep.TracedRun

SemanticModels.Parsers.AbstractCollector

SemanticModels.Parsers.FuncCollector

SemanticModels.Parsers.MetaCollector

SemanticModels.SpringModel

DiffEqBase.solve

SemanticModels.BasicSIR

SemanticModels.Dubstep.replacefunc

SemanticModels.Dubstep.replacenorm

SemanticModels.Dubstep.trace

SemanticModels.Graphs.copy_input_graph_to_new_graph

SemanticModels.Graphs.gen_rand_vertex_name

SemanticModels.Graphs.gen_vertex_hash

SemanticModels.Graphs.generate_synthetic_edges

SemanticModels.Graphs.generate_synthetic_vertices

SemanticModels.Graphs.insert_edges_from_jl

SemanticModels.Graphs.insert_edges_from_jl

SemanticModels.Graphs.insert_vertices_from_jl

SemanticModels.Graphs.insert_vertices_from_jl

SemanticModels.Graphs.load_graph_data

SemanticModels.Parsers.defs

SemanticModels.Parsers.funcs

SemanticModels.Parsers.parsefile

SemanticModels.hookeslaw